[Front] Implement EntityDetails component
